requirements.txt
aiogram
pyrogram
sqlalchemy
pydantic
aiosqlite
python-dotenv
alembic